rpk/metrics: Change free mem & free space to float in rpk, metrics & db

+++ src/go/rpk/pkg/cli/cmd/status.go
@@ -81,7 +81,7 @@ func executeStatus(
 			}
 		}
 		payload := api.MetricsPayload{
-			uint64(metrics.CpuPercentage),
+			metrics.CpuPercentage,
 			metrics.FreeMemoryMB,
 			metrics.FreeSpaceMB,
 		}
